pyproject.toml
[project]
name = "nutrilens-ai"
version = "1.0.0"
description = "Gemini-powered contextual nutrition intelligence platform"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0",
    "google-generativeai>=0.8.0",
    "firebase-admin>=6.5.0",
    "gspread>=6.0.0",
    "slowapi>=0.1.9",
    "httpx>=0.27.0",
    "Pillow>=10.0.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["services", "routers", "models", "config"]
